move multichannel.go to license package
