+++ pkg/kotsutil/kots.go
@@ -1627,7 +1627,10 @@ func FindChannelIDInLicense(requestedSlug string, license *kotsv1beta1.License) 
 }
 
 func FindChannelInLicense(channelID string, license *kotsv1beta1.License) (*kotsv1beta1.Channel, error) {
-	if channelID == "" || len(license.Spec.Channels) == 0 {
+	if channelID == "" {
+		return nil, errors.New("channelID is required")
+	}
+	if len(license.Spec.Channels) == 0 {
 		if license.Spec.ChannelID != channelID {
 			return nil, errors.New("channel not found in non-multi channel license")
 		}
